refactor: add animation effect for card in style, rename file for Cards
