+++ src/components/results/Cards.css
@@ -24,6 +24,43 @@
     box-shadow: inset 0 0 10px rgb(85, 255, 0);
 }
 
+.container-img_card {
+    position: relative;
+    width: 250px;
+    height: 350px;
+    perspective: 1000px;
+}
+
+.card-inner {
+    width: 100%;
+    height: 100%;
+    position: relative;
+    transform-style: preserve-3d;
+    transition: transform 0.6s ease;
+}
+
+.container-img_card:hover .card-inner {
+    transform: rotateY(180deg);
+}
+
+.card-front,
+.card-back {
+    position: absolute;
+    width: 100%;
+    height: 100%;
+    backface-visibility: hidden;
+}
+
+.card-front img,
+.card-back img {
+    width: 100%;
+    height: 100%;
+}
+
+.card-back {
+    transform: rotateY(180deg);
+}
+
 .main-section {
     display: flex;
     width: 95vw;
